+++ Interface/IUsuarioRepository.cs
@@ -5,7 +5,7 @@ namespace Clinica.Interface
     public interface IUsuarioRepository
     {
         Task<IEnumerable<Usuario>> GetUsuariosAsync();
-        Task<Usuario> GetByIdAsync(long id);
+        Task<Usuario> GetByIdAsync(int id);
         Task<Usuario> AddAsync(Usuario usuario);
         Task<Usuario> UpdateAsync(Usuario usuario);
         Task<Usuario> RemoveAsync(Usuario usuario);
