+++ Interface/IUsuarioService.cs
@@ -5,9 +5,9 @@ namespace Clinica.Interface
     public interface IUsuarioService
     {
         Task<IEnumerable<UsuarioDTO>> GetUsuarios();
-        Task<UsuarioDTO> GetById(long id);
+        Task<UsuarioDTO> GetById(int id);
         Task Add(UsuarioDTO usuarioDTO);
         Task Update(UsuarioDTO usuarioDTO);
-        Task Remove(long id);
+        Task Remove(int id);
     }
 }
