+++ Migrations/20230718020820_InitialMigration.cs
@@ -14,7 +14,7 @@ namespace Clinica.Migrations
                 name: "Medico",
                 columns: table => new
                 {
-                    Id = table.Column<long>(type: "bigint", nullable: false)
+                    Id = table.Column<int>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Nome = table.Column<string>(type: "varchar(20)", maxLength: 20, nullable: false),
                     CRM = table.Column<int>(type: "int", maxLength: 5, nullable: false),
